requirements.txt
Django==4.2.7
djangorestframework==3.14.0
drf-spectacular==0.26.5
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
openai==1.55.3
httpx==0.27.2
python-dotenv==1.0.0
gunicorn==21.2.0